+++ examples/BookAndChange/Program.cs
@@ -108,7 +108,7 @@ try
     var orderChange = await client.OrderChanges.Create(orderChangeOffers.Data.First().Id);
     Console.WriteLine($"Created order change with ID: {orderChange.Id}, confirming...");
 
-    var confirmedOrderChange = await client.OrderChanges.ConfirmPending(orderChange.Id, new Balance
+    var confirmedOrderChange = await client.OrderChanges.Confirm(orderChange.Id, new Balance
     {
         Amount = orderChange.ChangeTotalAmount,
         Currency = orderChange.ChangeTotalCurrency
